# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1",
    "paths": {
        "/": {
            "get": {
                "description": "redirect to root",
                "tags": [
                    "root"
                ],
                "summary": "redirect to root",
                "responses": {}
            }
        },
        "/dns": {
            "get": {
                "description": "GET all rows from SQLite for DNS data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dns"
                ],
                "summary": "GET DNS data from database",
                "responses": {
                    "200": {
                        "description": "Successfully extracted vulnerable packages from DNS",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DbPackage"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/domain": {
            "post": {
                "description": "Get a list of vulnerable packages founded in site and gau tool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gau"
                ],
                "summary": "Get vulnerable packages from site and gau",
                "parameters": [
                    {
                        "description": "Object for search",
                        "name": "inputGau",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GauStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of vulnerable packages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PackageManager"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/github/org": {
            "post": {
                "description": "Get dependency confusion packages that was founded in organization git repositories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org"
                ],
                "summary": "Get organization git repositories",
                "parameters": [
                    {
                        "description": "Get org repository packages",
                        "name": "org",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully extracted vulnerable packages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PackageManager"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/github/repo": {
            "post": {
                "description": "Get dependency confusion packages that was founded in git repositories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repo"
                ],
                "summary": "Get a git repository",
                "parameters": [
                    {
                        "description": "Get repository packages",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully extracted vulnerable packages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PackageManager"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/parse/directory": {
            "post": {
                "description": "Parse a directory for dependency confusion",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dir"
                ],
                "summary": "Parse a directory for dependency confusion",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Files for scanning",
                        "name": "files[]",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully extracted vulnerable packages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PackageManager"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/parse/file": {
            "post": {
                "description": "Parse a file for dependency confusion",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "Parse a file for dependency confusion",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File for scanning",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully extracted vulnerable packages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PackageManager"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/registry": {
            "post": {
                "description": "POST vulnerable version of package to corresponding registry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registry"
                ],
                "summary": "POST data to registry",
                "parameters": [
                    {
                        "description": "Package object",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PackageManager"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DataExfiltrated": {
            "type": "object",
            "properties": {
                "c": {
                    "type": "string"
                },
                "d": {
                    "type": "string"
                },
                "h": {
                    "type": "string"
                },
                "p": {
                    "type": "string"
                }
            }
        },
        "models.DbPackage": {
            "type": "object",
            "properties": {
                "dataExfiltrated": {
                    "$ref": "#/definitions/models.DataExfiltrated"
                },
                "id": {
                    "type": "integer"
                },
                "sourceIP": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.GauStruct": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "retries": {
                    "type": "integer"
                },
                "threads": {
                    "type": "integer"
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "models.PackageManager": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "package": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}